删除挑战的地方加上Modal确认, 学生的header菜单显示班级信息

+++ src/pages/admin/AdminBoardManage.tsx
@@ -1,9 +1,10 @@
-import { useEffect, useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import Sidebar from '@/partials/Sidebar'
 import Header from '@/partials/Header'
 import { baseURL } from '@/utils/Utils.ts'
 import { ChessBoardInfo } from '@/types/Chess.ts'
 import { useNavigate } from 'react-router-dom'
+import ModalBlank from '@/components/ModalBlank'
 
 function ChessboardManagement() {
   const navigate = useNavigate()
@@ -32,9 +33,13 @@ function ChessboardManagement() {
   }
 
   // 删除挑战
-  const handleDeleteChessboard = async (chessboardId: number) => {
+  const handleDeleteChessboard = async () => {
+    if (!currentBoard.current) {
+      alert('未选择挑战')
+      return
+    }
     const response = await fetch(
-      `${baseURL}/chessboard/${chessboardId}`,
+      `${baseURL}/chessboard/${currentBoard.current.id}`,
       {
         method: 'DELETE'
       }
@@ -42,6 +47,7 @@ function ChessboardManagement() {
     if (response.ok) {
       alert('挑战已成功删除')
       await fetchChessboard() // 更新挑战列表
+      setOpenDeleteBoard(false)
     } else {
       alert('删除挑战失败')
     }
@@ -51,6 +57,10 @@ function ChessboardManagement() {
   const viewChessboard = (chessboardId: number) => {
     navigate(`/admin/chessboard/${chessboardId}`)
   }
+
+  // 删除挑战的Modal
+  const [openDeleteBoard, setOpenDeleteBoard] = useState<boolean>(false)
+  const currentBoard = useRef<ChessBoardInfo>()
 
   useEffect(() => {
     fetchChessboard().then()
@@ -87,9 +97,9 @@ function ChessboardManagement() {
                     </tr>
                   </thead>
                   <tbody>
-                    {chessboardList.map((chessboard) => (
+                    {chessboardList.map((chessboard, index) => (
                       <tr key={chessboard.id}>
-                        <td className="border px-4 py-2">{chessboard.id}</td>
+                        <td className="border px-4 py-2">{index + 1}</td>
                         <td className="border px-4 py-2">{chessboard.name}</td>
                         <td className="border px-4 py-2">{chessboard.score}</td>
                         <td className="border px-4 py-2">
@@ -101,9 +111,12 @@ function ChessboardManagement() {
                             </button>
                             <button
                               className="btn bg-red-500 hover:bg-red-600 text-white"
-                              onClick={() =>
-                                handleDeleteChessboard(chessboard.id)
-                              }
+                              aria-controls="modal-delete-board"
+                              onClick={e => {
+                                e.stopPropagation()
+                                currentBoard.current = chessboard
+                                setOpenDeleteBoard(true)
+                              }}
                             >删除
                             </button>
                           </div>
@@ -118,6 +131,35 @@ function ChessboardManagement() {
             </div>
           </div>
         </main>
+        {/* 删除挑战的Modal */}
+        <ModalBlank id="modal-delete-board" modalOpen={openDeleteBoard} setModalOpen={setOpenDeleteBoard}>
+          <div className="p-5 flex space-x-4">
+            {/* Icon */}
+            <div className="w-10 h-10 rounded-full flex items-center justify-center shrink-0 bg-rose-100 dark:bg-rose-500/30">
+              <svg className="w-4 h-4 shrink-0 fill-current text-rose-500" viewBox="0 0 16 16">
+                <path d="M8 0C3.6 0 0 3.6 0 8s3.6 8 8 8 8-3.6 8-8-3.6-8-8-8zm0 12c-.6 0-1-.4-1-1s.4-1 1-1 1 .4 1 1-.4 1-1 1zm1-3H7V4h2v5z" />
+              </svg>
+            </div>
+            {/* Content */}
+            <div className="flex-1">
+              {/* Modal header */}
+              <div className="mb-2">
+                <div className="text-lg font-semibold text-slate-800 dark:text-slate-100">删除挑战</div>
+              </div>
+              {/* Modal content */}
+              <div className="text-sm mb-10">
+                <div className="space-y-2">
+                  <p>是否确认删除该挑战?</p>
+                </div>
+              </div>
+              {/* Modal footer */}
+              <div className="flex flex-wrap justify-end space-x-2">
+                <button className="btn-sm border-slate-200 dark:border-slate-700 hover:border-slate-300 dark:hover:border-slate-600 text-slate-600 dark:text-slate-300" onClick={(e) => { e.stopPropagation(); setOpenDeleteBoard(false); }}>取消</button>
+                <button onClick={handleDeleteChessboard} className="btn-sm bg-rose-500 hover:bg-rose-600 text-white">确认删除</button>
+              </div>
+            </div>
+          </div>
+        </ModalBlank>
       </div>
     </div>
   )
