+++ src/pages/teacher/ChessboardManagement.tsx
@@ -102,16 +102,22 @@ function ChessboardManagement() {
     }
   }
 
-  // 删除挑战
-  const handleDeleteChessboard = async (chessboardId: number) => {
+  // 直接删除挑战
+  const [openDeleteBoard, setOpenDeleteBoard] = useState<boolean>(false)
+  const handleDeleteChessboard = async () => {
+    if (!selectedBoard) {
+      alert('未选择挑战')
+      return
+    }
     const response = await fetch(
-      `${baseURL}/chessboard/${chessboardId}`,
+      `${baseURL}/chessboard/${selectedBoard.id}`,
       {
         method: 'DELETE'
       }
     )
     if (response.ok) {
       alert('挑战已成功删除')
+      setOpenDeleteBoard(false)
       await fetchChessboard() // 更新挑战列表
     } else {
       alert('删除挑战失败')
@@ -119,7 +125,7 @@ function ChessboardManagement() {
   }
 
   // 从班级删除挑战
-  const [openDeleteBoard, setOpenDeleteBoard] = useState<boolean>(false)
+  const [openDeleteBoardFromClass, setOpenDeleteBoardFromClass] = useState<boolean>(false)
   const deleteBoardFromClass = async () => {
     if (!selectedBoard || !selectedClass) {
       alert('未选中任何残局')
@@ -132,7 +138,7 @@ function ChessboardManagement() {
     if (res.ok) {
       alert('已从班级删除对应挑战')
       await fetchChessboardByClass()
-      setOpenDeleteBoard(false)
+      setOpenDeleteBoardFromClass(false)
     } else {
       alert('删除失败')
     }
@@ -219,7 +225,7 @@ function ChessboardManagement() {
                               onClick={e => {
                                 e.stopPropagation()
                                 setSelectedBoard(chessboard)
-                                setOpenDeleteBoard(true)
+                                setOpenDeleteBoardFromClass(true)
                               }}
                             >删除
                             </button>
@@ -269,9 +275,12 @@ function ChessboardManagement() {
                             </button>
                             <button
                               className="btn bg-red-500 hover:bg-red-600 text-white"
-                              onClick={() =>
-                                handleDeleteChessboard(chessboard.id)
-                              }
+                              aria-controls="modal-delete-board"
+                              onClick={e => {
+                                e.stopPropagation()
+                                setSelectedBoard(chessboard)
+                                setOpenDeleteBoard(true)
+                              }}
                             >删除
                             </button>
                           </div>
@@ -319,7 +328,36 @@ function ChessboardManagement() {
           </div>
         </ModalBasic>
         {/* 从班级删除挑战的Modal */}
-        <ModalBlank id="modal-delete-board-fromClass" modalOpen={openDeleteBoard} setModalOpen={setOpenDeleteBoard}>
+        <ModalBlank id="modal-delete-board-fromClass" modalOpen={openDeleteBoardFromClass} setModalOpen={setOpenDeleteBoardFromClass}>
+          <div className="p-5 flex space-x-4">
+            {/* Icon */}
+            <div className="w-10 h-10 rounded-full flex items-center justify-center shrink-0 bg-rose-100 dark:bg-rose-500/30">
+              <svg className="w-4 h-4 shrink-0 fill-current text-rose-500" viewBox="0 0 16 16">
+                <path d="M8 0C3.6 0 0 3.6 0 8s3.6 8 8 8 8-3.6 8-8-3.6-8-8-8zm0 12c-.6 0-1-.4-1-1s.4-1 1-1 1 .4 1 1-.4 1-1 1zm1-3H7V4h2v5z" />
+              </svg>
+            </div>
+            {/* Content */}
+            <div className="flex-1">
+              {/* Modal header */}
+              <div className="mb-2">
+                <div className="text-lg font-semibold text-slate-800 dark:text-slate-100">删除班级挑战</div>
+              </div>
+              {/* Modal content */}
+              <div className="text-sm mb-10">
+                <div className="space-y-2">
+                  <p>是否确认要从班级删除该挑战?</p>
+                </div>
+              </div>
+              {/* Modal footer */}
+              <div className="flex flex-wrap justify-end space-x-2">
+                <button className="btn-sm border-slate-200 dark:border-slate-700 hover:border-slate-300 dark:hover:border-slate-600 text-slate-600 dark:text-slate-300" onClick={(e) => { e.stopPropagation(); setOpenDeleteBoardFromClass(false); }}>取消</button>
+                <button onClick={deleteBoardFromClass} className="btn-sm bg-rose-500 hover:bg-rose-600 text-white">确认删除</button>
+              </div>
+            </div>
+          </div>
+        </ModalBlank>
+        {/* 直接删除挑战的Modal */}
+        <ModalBlank id="modal-delete-board" modalOpen={openDeleteBoard} setModalOpen={setOpenDeleteBoard}>
           <div className="p-5 flex space-x-4">
             {/* Icon */}
             <div className="w-10 h-10 rounded-full flex items-center justify-center shrink-0 bg-rose-100 dark:bg-rose-500/30">
@@ -336,13 +374,13 @@ function ChessboardManagement() {
               {/* Modal content */}
               <div className="text-sm mb-10">
                 <div className="space-y-2">
-                  <p>是否确认要从班级删除该挑战?</p>
+                  <p>是否确认删除该挑战?</p>
                 </div>
               </div>
               {/* Modal footer */}
               <div className="flex flex-wrap justify-end space-x-2">
                 <button className="btn-sm border-slate-200 dark:border-slate-700 hover:border-slate-300 dark:hover:border-slate-600 text-slate-600 dark:text-slate-300" onClick={(e) => { e.stopPropagation(); setOpenDeleteBoard(false); }}>取消</button>
-                <button onClick={deleteBoardFromClass} className="btn-sm bg-rose-500 hover:bg-rose-600 text-white">确认删除</button>
+                <button onClick={handleDeleteChessboard} className="btn-sm bg-rose-500 hover:bg-rose-600 text-white">确认删除</button>
               </div>
             </div>
           </div>
